+++ packages/extension-polkagate/src/fullscreen/governance/delegate/partial/AmountWithOptionsAndLockAmount.tsx
@@ -1,19 +1,19 @@
 // Copyright 2019-2024 @polkadot/extension-polkagate authors & contributors
 // SPDX-License-Identifier: Apache-2.0
-// @ts-nocheck
 
 /* eslint-disable react/jsx-max-props-per-line */
 
+import type { BN } from '@polkadot/util';
+import type { Lock } from '../../../../hooks/useAccountLocks';
 import type { BalancesInfo } from '../../../../util/types';
 
 import { Grid, Typography } from '@mui/material';
-import React, { useMemo } from 'react';
+import React from 'react';
 
-import { BN, isBn } from '@polkadot/util';
+import { isBn } from '@polkadot/util';
 
 import { AmountWithOptions, Infotip2, ShowBalance } from '../../../../components';
 import { useTranslation } from '../../../../hooks';
-import { Lock } from '../../../../hooks/useAccountLocks';
 import { amountToHuman } from '../../../../util/utils';
 import { getLockedUntil } from '../../utils/util';
 
@@ -30,11 +30,18 @@ interface Props {
   amount?: string | undefined;
 }
 
-export default function AmountWithOptionsAndLockAmount({ accountLocks, amount, balances, currentBlock, decimal, lockedAmount, onLockedAmount, onMaxAmount, onValueChange, token }: Props): React.ReactElement {
+interface AlreadyLockedTooltipTextProps {
+  accountLocks: Lock[];
+  currentBlock: number;
+  decimal: number | undefined;
+  token: string | undefined;
+}
+
+const AlreadyLockedTooltipText = ({ accountLocks, currentBlock, decimal, token }: AlreadyLockedTooltipTextProps) => {
   const { t } = useTranslation();
 
-  const alreadyLockedTooltipText = useMemo(() => accountLocks && currentBlock &&
-    (<Grid container item sx={{ maxHeight: '400px', overflow: 'hidden', overflowY: 'scroll' }}>
+  return (
+    <Grid container item sx={{ maxHeight: '400px', overflow: 'hidden', overflowY: 'scroll' }}>
       <Typography variant='body2'>
         <Grid container spacing={2}>
           <Grid item xs={2.5}>
@@ -68,9 +75,22 @@ export default function AmountWithOptionsAndLockAmount({ accountLocks, amount, b
         </Grid>
       </Typography>
     </Grid>
-    ), [accountLocks, currentBlock, decimal, t, token]);
+  );
+};
 
-  const VotingInformation = () => (
+interface VotingInformationProps {
+  balances: BalancesInfo | undefined;
+  decimal: number | undefined;
+  token: string | undefined;
+  lockedAmount: BN | undefined;
+  accountLocks: Lock[] | null | undefined;
+  currentBlock: number | undefined;
+}
+
+const VotingInformation = ({ accountLocks, balances, currentBlock, decimal, lockedAmount, token }: VotingInformationProps) => {
+  const { t } = useTranslation();
+
+  return (
     <Grid container item pb='10px'>
       <Grid container item justifyContent='space-between' sx={{ mt: '10px', width: '70.25%' }}>
         <Grid item sx={{ fontSize: '16px', fontWeight: 400 }}>
@@ -81,31 +101,41 @@ export default function AmountWithOptionsAndLockAmount({ accountLocks, amount, b
         </Grid>
       </Grid>
       {lockedAmount && !lockedAmount?.isZero() &&
-        <Grid alignItems='center' container item justifyContent='space-between' sx={{ lineHeight: '20px', width: '75%' }}>
-          <Grid item sx={{ fontSize: '16px', fontWeight: 400 }}>
-            <Infotip2 showQuestionMark text={t('The maximum number of tokens that are already locked in the ecosystem')}>
-              {t('Already Locked Amount')}
-            </Infotip2>
-          </Grid>
-          <Grid item sx={{ fontSize: '16px', fontWeight: 500 }}>
-            <Infotip2 showInfoMark text={alreadyLockedTooltipText || t('Fetching ...')}>
-              <ShowBalance balance={lockedAmount} decimal={decimal} decimalPoint={2} token={token} />
-            </Infotip2>
-          </Grid>
+      <Grid alignItems='center' container item justifyContent='space-between' sx={{ lineHeight: '20px', width: '75%' }}>
+        <Grid item sx={{ fontSize: '16px', fontWeight: 400 }}>
+          <Infotip2 showQuestionMark text={t('The maximum number of tokens that are already locked in the ecosystem')}>
+            {t('Already Locked Amount')}
+          </Infotip2>
         </Grid>
+        <Grid item sx={{ fontSize: '16px', fontWeight: 500 }}>
+          <Infotip2
+            showInfoMark
+            text={accountLocks && currentBlock
+              ? <AlreadyLockedTooltipText accountLocks={accountLocks} currentBlock={currentBlock} decimal={decimal} token={token} />
+              : t('Fetching ...')
+            }
+          >
+            <ShowBalance balance={lockedAmount} decimal={decimal} decimalPoint={2} token={token} />
+          </Infotip2>
+        </Grid>
+      </Grid>
       }
     </Grid>
   );
+};
+
+export default function AmountWithOptionsAndLockAmount ({ accountLocks, amount, balances, currentBlock, decimal, lockedAmount, onLockedAmount, onMaxAmount, onValueChange, token }: Props): React.ReactElement {
+  const { t } = useTranslation();
 
   return (
     <>
       <AmountWithOptions
-        label={t<string>('Amount ({{token}})', { replace: { token } })}
+        label={t('Amount ({{token}})', { replace: { token } })}
         onChangeAmount={onValueChange}
         onPrimary={onMaxAmount}
-        onSecondary={onLockedAmount}
-        primaryBtnText={t<string>('Max amount')}
-        secondaryBtnText={t<string>('Locked amount')}
+        onSecondary={lockedAmount && !lockedAmount?.isZero() ? onLockedAmount : undefined}
+        primaryBtnText={t('Max amount')}
+        secondaryBtnText={t('Locked amount')}
         style={{
           '> div div div': { fontSize: '16px', fontWeight: 400 },
           fontSize: '16px',
@@ -114,7 +144,14 @@ export default function AmountWithOptionsAndLockAmount({ accountLocks, amount, b
         }}
         value={amount}
       />
-      <VotingInformation />
+      <VotingInformation
+        accountLocks={accountLocks}
+        balances={balances}
+        currentBlock={currentBlock}
+        decimal={decimal}
+        lockedAmount={lockedAmount}
+        token={token}
+      />
     </>
   );
 }